wen-Version Estable

+++ td/web/WEB-INF/jsp/modificarAnuncio.jsp
@@ -11,6 +11,7 @@
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <%@include file="../../interfaz/inicio.html" %>
 <!DOCTYPE html>
+<script type="text/javascript" src="interfaz/verificaciones.js"></script>
 <div id="content"><div class="inner_copy"><a href="http://www.freetemplatesonline.com/">Free Web Templates</a> <a href="http://www.websitetemplates.org/">Website Templates</a> <a href="http://www.webdesign.org/website-design">Website Design</a></div>
     <div class="post">
 	
@@ -20,11 +21,13 @@
             <h2 class="title">Modificar Anuncio:</h2>
             <br/>
         
-             ${error}
-             ${mensaje}
+             <script>
+             Mensaje("${mensaje}");
+             Mensaje("${error}")
+                </script>
 
         
-         <a href="eliminarAnuncio.htm?id=${p.codanuncio}" >Eliminar</a> 
+         <a href="eliminarAnuncio.htm?id=${p.codanuncio}" onclick="return confirmar('¿Está seguro que desea eliminar este anuncio?')">Eliminar</a> 
         <form:form action="modificarAnuncio.htm?id=${p.codanuncio}" commandName="ma" >
             <table>
                 <tr>
